[flang] Expose specific to generic intrinsic name mapping
The intrinsic lowering facility is based on the generic intrinsics to avoid
duplicating implementations. Specific intrinsics call are re-written to call to
the generic versions by the front-end but this is cannot done when specific intrinsics
are passed as arguments (the rewrite would yield illegal/ambiguous unparsed Fortran).
Solve the issue by making the specific to generic name mapping accessible in lowering
and so that it can be used to generate the unrestricted intrinsic functions.