Rework earlier fix to stringification in the preprocessor
An earlier change fixed the preprocessing of the #x stringification
operation (if that's a word) by forcing macro replacement in the
argument beforehand.  This turned out to fix the original problem
case by cancelling the effects of another bug.  A better reading of
the C17 standard is to perform macro replacement of arguments to
function-like macros unless they're being stringified or pasted.